+++ Source/Scene/Globe.js
@@ -251,6 +251,15 @@ import TileSelectionResult from './TileSelectionResult.js';
         this.atmosphereBrightnessShift = 0.0;
 
         /**
+         * Whether to show terrain skirts. Terrain skirts are geometry extending downwards from a tile's edges used to hide seams between neighboring tiles.
+         * It may be desirable to hide terrain skirts if terrain is translucent or when viewing terrain from below the surface.
+         *
+         * @type {Boolean}
+         * @default true
+         */
+        this.showSkirts = true;
+
+        /**
          * Whether to cull back-facing terrain. Set this to false when viewing terrain from below the surface.
          *
          * @type {Boolean}
@@ -754,6 +763,7 @@ import TileSelectionResult from './TileSelectionResult.js';
             tileProvider.saturationShift = this.atmosphereSaturationShift;
             tileProvider.brightnessShift = this.atmosphereBrightnessShift;
             tileProvider.fillHighlightColor = this.fillHighlightColor;
+            tileProvider.showSkirts = this.showSkirts;
             tileProvider.backFaceCulling = this.backFaceCulling;
             surface.beginFrame(frameState);
         }